Addresses review feedback, including removing response
