+++ extensions/amp-ad-network-fake-impl/0.1/data/fake_amp_ad_with_3p_analytics.html
@@ -8,7 +8,7 @@
     <meta name=viewport content=width=device-width,minimum-scale=1>
   </head>
   <body>
-    <a href="/landing.html?testVendorStatus=3PANALYTICS(testVendor,status)&testVendorOther=3PANALYTICS(testVendor,other)&testVendor2Status=3PANALYTICS(testVendor2,status)&testVendor2Other=3PANALYTICS(testVendor2,other)">
+    <a href="/landing.html">
       <amp-img id=img src='https://upload.wikimedia.org/wikipedia/commons/6/6e/Golde33443.jpg' width=256 height=300></amp-img>
       <p class=frob>
         By Golden Trvs Gol twister (Own work) [CC BY-SA 4.0 (http://creativecommons.org/licenses/by-sa/4.0)], via Wikimedia Commons
