+++ src/3p-analytics-common.js
@@ -17,7 +17,6 @@
 /** @enum {string} */
 export const AMP_ANALYTICS_3P_MESSAGE_TYPE = {
   EVENT: 'E',
-  RESPONSE: 'R',
 };
 
 /** @typedef {!Object<!string,!Array<!string>>} */
@@ -28,56 +27,3 @@ export let AmpAnalytics3pEvent;
 //   ...
 // }
 
-/** @typedef {JsonObject} */
-export let AmpAnalytics3pResponse;
-// Example:
-// {"status":"received","somethingElse":"42"}
-
-/**
- * A class for holding AMP Analytics third-party vendors responses to frames.
- */
-export class ResponseMap {
-  /**
-   * Add a response. Note that only the most recent one for a given
-   * ampDoc+frameType+creativeUrl is stored. If add() is called again with
-   * the same first three parameters, the old value will be overwritten.
-   * @param {!string} frameType The identifier for the third-party frame that
-   * responded
-   * @param {!string} creativeUrl The URL of the creative being responded to
-   * @param {Object} response What the response was
-   */
-  static add(ampDoc, frameType, creativeUrl, response) {
-    const map = ampDoc.getAnchorClickListenerBinding();
-    map[frameType] = map[frameType] || {};
-    map[frameType][creativeUrl] = response;
-  }
-
-  /**
-   * Gets the most recent response given by a certain frame to a certain
-   * creative
-   * @param {!string} frameType The identifier for the third-party frame
-   * whose response is sought
-   * @param {!string} creativeUrl The URL of the creative that the sought
-   * response was about
-   * @returns {?Object}
-   */
-  static get(ampDoc, frameType, creativeUrl) {
-    const map = ampDoc.getAnchorClickListenerBinding();
-    if (map[frameType] && map[frameType][creativeUrl]) {
-      return map[frameType][creativeUrl];
-    }
-    return {};
-  }
-
-  /**
-   * Remove a response, for instance if a third-party frame is being destroyed
-   * @param {!string} frameType The identifier for the third-party frame
-   * whose responses are to be removed
-   */
-  static remove(ampDoc, frameType) {
-    const map = ampDoc.getAnchorClickListenerBinding();
-    if (map[frameType]) {
-      delete map[frameType];
-    }
-  }
-}
